third_party/blink/web_tests/external/wpt/webaudio/the-audio-api/the-mediaelementaudiosourcenode-interface/no-cors.https.html

<!DOCTYPE html>
<html>
  <head>
    <title>
      Test if MediaElementAudioSourceNode works for cross-origin redirects with
      "no-cors" request mode.
    </title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/common/get-host-info.sub.js"></script>
  </head>
  <body>
    <script>
      const filePath = '/webaudio/js/worklet-recorder.js';

      promise_test(async () => {
        const context = new AudioContext();
        await context.suspend();

        const host_info = get_host_info();
        const audioElement = document.createElement('audio');
        audioElement.loop = true;
        const wav =
            `${host_info.HTTPS_ORIGIN}/webaudio/resources/4ch-440.wav?` +
            `pipe=header(access-control-allow-origin,*)`;
        audioElement.src =
            `${host_info.HTTPS_REMOTE_ORIGIN}` +
            `/fetch/api/resources/redirect.py?location=` +
            encodeURIComponent(wav);

        await context.audioWorklet.addModule(filePath);

        const source = new MediaElementAudioSourceNode(context, {
          mediaElement: audioElement,
        });
        const workletRecorder = new AudioWorkletNode(
            context, 'recorder-processor', {channelCount: 4});
        source.connect(workletRecorder).connect(context.destination);

        const recordingPromise = new Promise((resolve, reject) => {
          workletRecorder.port.onmessage = (event) => {
            if (event.data && event.data.type === 'recordfinished') {
              resolve(event.data.recordBuffer);
            }
          };
        });

        await context.resume();
        await audioElement.play();

        const recordBuffer = await recordingPromise;

        // The recorded data from MESN must be zero.
        for (let i = 0; i < recordBuffer.length; ++i) {
          const channelData = recordBuffer[i];
          const expected = new Float32Array(channelData.length);
          assert_array_equals(
              channelData, expected, `Recorded channel #${i}`);
        }
      }, 'cross-origin redirect with no-cors');
    </script>
  </body>
</html>
